requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
tesseract-ocr==1.0.0
openai==1.3.0
python-magic==0.4.27
boto3==1.29.0
psycopg2-binary==2.9.9
motor==3.5.0
python-jose[cryptography]==3.3.0
fastapi-jwt-auth==0.5.0
python-magic-bin==0.4.14
alembic==1.12.1
sqlalchemy==2.0.23
python-magic==0.4.27

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring
prometheus-fastapi-instrumentator==5.10.0
sentry-sdk==1.36.0

# Analytics
pandas==2.1.0
numpy==1.24.3

# Frontend
fastapi-staticfiles==0.2.0
jinja2==3.1.2

# Integration
websockets==11.0.3
aiohttp==3.9.1

# Document Processing
pdf2image==1.16.3
python-docx==0.8.11

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.25.1
